+++ R/FitRsmRoc.R
@@ -11,7 +11,7 @@
 #' @return The return value is a list with the following elements:
 #' @return \item{mu}{The mean of the diseased distribution relative 
 #'    to the non-diseased one} 
-#' @return \item{lambdP}{The Poisson parameter describing the distribution 
+#' @return \item{lambda}{The Poisson parameter describing the distribution 
 #'    of latent NLs per case}
 #' @return \item{nu}{The binomial success probability describing the distribution
 #'    of latent LLs per diseased case}
@@ -219,7 +219,7 @@ FitRsmRoc <- function(binnedRocData, lesDistr, trt = 1, rdr = 1){
   zetas <- InverseZetas(ret@coef[4:length(ret@coef)])
   
   # 11/30/20
-  retx <- UtilRSM2IntrinsicRSM(mu, lambda, nu)
+  retx <- UtilRSM2Intrinsic(mu, lambda, nu)
   lambda_i <- retx$lambda_i
   nu_i <- retx$nu_i