+++ tests/testthat/test-predicted-plots.R
@@ -1,7 +1,7 @@
 contextStr <- "PlotBinormalFit, PlotCbmFit, PlotRsmFit"
 context(contextStr)
 test_that(contextStr, {
-
+  
   fn <- paste0(test_path(), "/goodValues361/Plots/Binormal", ".rds")
   if (!file.exists(fn)) {
     warning(paste0("File not found - generating new ",fn))
@@ -37,6 +37,21 @@ test_that("PlotCbmFit", {
 
 
 
+UndoSearchParameterChange <- function (mu, lambda_i, nu_i)
+{
+  lambda <- lambda_i * 0
+  nu <- nu_i * 0
+  for (i in 1:length(lambda)) {
+    x <- UtilIntrinsic2RSM(mu[i], lambda_i[i], nu_i[i])
+    lambda[i] <- x$lambda
+    nu[i] <- x$nu
+  }
+  return (list(
+    lambda = lambda,
+    nu = nu
+  ))
+}
+
 
 test_that("Rsm1", {
   Sys.sleep(0.2)
@@ -46,6 +61,10 @@ test_that("Rsm1", {
   nu_i <- c(0.6, 0.8) 
   zeta1 <- c(-3,-3) 
   
+  x <- UndoSearchParameterChange(mu, lambda_i, nu_i)
+  lambda <- x$lambda
+  nu <- x$nu
+  
   lesDistr <- c(0.2, 0.4, 0.1, 0.3)
   relWeights <- c(0.3, 0.4, 0.2,  0.1)
   
@@ -53,8 +72,8 @@ test_that("Rsm1", {
   if (!file.exists(fn)) {
     warning(paste0("File not found - generating new ",fn))
     ret <- PlotRsmOperatingCharacteristics(mu = mu, 
-                                           lambda_i = lambda_i, 
-                                           nu_i = nu_i, 
+                                           lambda = lambda, 
+                                           nu = nu, 
                                            zeta1 = zeta1, 
                                            OpChType = "wAFROC",
                                            lesDistr = lesDistr, 
@@ -64,13 +83,13 @@ test_that("Rsm1", {
   
   ret <- readRDS(fn)
   expect_equal(PlotRsmOperatingCharacteristics(mu = mu, 
-                                               lambda_i = lambda_i, 
-                                               nu_i = nu_i, 
+                                               lambda = lambda, 
+                                               nu = nu, 
                                                zeta1 = zeta1, 
                                                OpChType = "wAFROC",
                                                lesDistr = lesDistr, 
                                                relWeights = relWeights), ret, check.environment = FALSE)
-
+  
 })
 
 
@@ -83,19 +102,23 @@ test_that("Rsm2", {
   for (i in nLesPerCase) lesDistr[i] <- sum(Lk2 == i)/K2
   
   maxLL <- max(lesDistr)
-
+  
   mu <- 10;nu_i <- 0.99;lambda_i <- 1
+  
+  x <- UndoSearchParameterChange(mu, lambda_i, nu_i)
+  lambda <- x$lambda
+  nu <- x$nu
   
   fn <- paste0(test_path(), "/goodValues361/Plots/Rsm2", ".rds")
   if (!file.exists(fn)) {
     warning(paste0("File not found - generating new ",fn))
-    ret <- PlotRsmOperatingCharacteristics(mu, lambda_i, nu_i, OpChType = "wAFROC", 
-                                            lesDistr = lesDistr)
+    ret <- PlotRsmOperatingCharacteristics(mu, lambda, nu, OpChType = "wAFROC", 
+                                           lesDistr = lesDistr)
     saveRDS(ret, file = fn)
   }
   
   ret <- readRDS(fn)
-  expect_equal(PlotRsmOperatingCharacteristics(mu, lambda_i, nu_i, OpChType = "wAFROC", 
+  expect_equal(PlotRsmOperatingCharacteristics(mu, lambda, nu, OpChType = "wAFROC", 
                                                lesDistr = lesDistr), ret, check.environment = FALSE)
   
 })
@@ -110,19 +133,26 @@ test_that("RSM3", {
                                           package = "RJafroc", mustWork = TRUE)
   x <- DfReadDataFile(frocCrAbnormalCasesFirst, newExcelFileFormat = TRUE)
   lesDistr <- UtilLesionDistrVector(x)
-
+  
+  mu <- c(2, 3) 
+  lambda_i <- c(1, 1.5)
+  nu_i <- c(0.6, 0.8) 
+  zeta1  <- c(-3,-3)
+  
+  x <- UndoSearchParameterChange(mu, lambda_i, nu_i)
+  lambda <- x$lambda
+  nu <- x$nu
+  
   fn <- paste0(test_path(), "/goodValues361/Plots/Rsm3", ".rds")
   if (!file.exists(fn)) {
     warning(paste0("File not found - generating new ",fn))
     
-    ret <- PlotRsmOperatingCharacteristics(mu = c(2, 3), lambda_i = c(1, 1.5), nu_i = c(0.6, 0.8),  zeta1 = c(-3,-3), OpChType = "wAFROC",
-                                           lesDistr = lesDistr, legendPosition = "bottom", nlfRange = c(0, 1), llfRange = c(0, 1))
+    ret <- PlotRsmOperatingCharacteristics(mu = mu, lambda = lambda, nu = nu,  zeta1 = zeta1, OpChType = "wAFROC", lesDistr = lesDistr, legendPosition = "bottom", nlfRange = c(0, 1), llfRange = c(0, 1))
     saveRDS(ret, file = fn)
   }
   
   ret <- readRDS(fn)
-  expect_equal(PlotRsmOperatingCharacteristics(mu = c(2, 3), lambda_i = c(1, 1.5), nu_i = c(0.6, 0.8), zeta1 = c(-3,-3), OpChType = "wAFROC",
-                                               lesDistr = lesDistr, legendPosition = "bottom", nlfRange = c(0, 1), llfRange = c(0, 1)), ret, check.environment = FALSE)
+  expect_equal(PlotRsmOperatingCharacteristics(mu = mu, lambda = lambda, nu = nu,  zeta1 = zeta1, OpChType = "wAFROC", lesDistr = lesDistr, legendPosition = "bottom", nlfRange = c(0, 1), llfRange = c(0, 1)), ret, check.environment = FALSE)
   
 })